+++ src/drivers/generators/BaseCacheGenerator.php
@@ -170,13 +170,15 @@ abstract class BaseCacheGenerator extends SavableComponent implements CacheGener
      * @param SiteUriModel[]|array[] $siteUris
      * @return array
      */
-    protected function getUrlsToGenerate(array $siteUris): array
+    protected function getUrlsToGenerate(array $siteUris, bool $withToken = true): array
     {
         $urls = [];
         $nonCacheableSiteUris = [];
-        $params = [
-            'token' => Craft::$app->getTokens()->createToken(self::GENERATE_ACTION_ROUTE),
-        ];
+        $params = [];
+
+        if ($withToken) {
+            $params['token'] = Craft::$app->getTokens()->createToken(self::GENERATE_ACTION_ROUTE);
+        }
 
         foreach ($siteUris as $siteUri) {
             // Convert to a site URI model if it is an array
@@ -218,14 +220,6 @@ abstract class BaseCacheGenerator extends SavableComponent implements CacheGener
         }
 
         return $count;
-    }
-
-    /**
-     * Returns whether the provided URL is a page (not an include).
-     */
-    protected function isPageUrl(string $url): bool
-    {
-        return !str_contains($url, CacheRequestService::CACHED_INCLUDE_PATH . '?action=' . CacheRequestService::CACHED_INCLUDE_ACTION);
     }
 
     protected function outputVerbose(string $url, bool $success = true): void