+++ src/drivers/storage/CacheStorageTrait.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+namespace putyourlightson\blitz\drivers\storage;
+
+trait CacheStorageTrait
+{
+    /**
+     * @var bool Whether this is a dummy storage drive.
+     */
+    public bool $isDummy = false;
+}