+++ src/enums/HeaderEnum.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+namespace putyourlightson\blitz\enums;
+
+/**
+ * TODO: convert to an Enum in Blitz 5.0
+ */
+class HeaderEnum
+{
+    public const ACCEPT_ENCODING = 'Accept-Encoding';
+    public const CACHE_CONTROL = 'Cache-Control';
+    public const CACHE_TAG = 'Cache-Tag';
+    public const CONTENT_ENCODING = 'Content-Encoding';
+    public const CONTENT_TYPE = 'Content-Type';
+    public const PERMISSIONS_POLICY = 'Permissions-Policy';
+    public const X_POWERED_BY = 'X-Powered-By';
+    public const X_ROBOTS_TAG = 'X-Robots-Tag';
+}