+++ tests/pest/Interface/WebResponseTest.php
@@ -6,11 +6,13 @@
 
 use Mockery\MockInterface;
 use putyourlightson\blitz\Blitz;
+use putyourlightson\blitz\enums\HeaderEnum;
 use putyourlightson\blitz\models\SettingsModel;
 use putyourlightson\blitz\services\RefreshCacheService;
 
 beforeEach(function() {
     Blitz::$plugin->cacheStorage->deleteAll();
+    Blitz::$plugin->generateCache->options->outputComments = null;
     Blitz::$plugin->set('refreshCache', Mockery::mock(RefreshCacheService::class . '[refresh]'));
 });
 
@@ -22,14 +24,14 @@ test('Response contains the default cache control header when the page is not ca
     $response = sendRequest();
     Blitz::$plugin->cacheRequest->setDefaultCacheControlHeader();
 
-    expect($response->headers->get('cache-control'))
+    expect($response->headers->get(HeaderEnum::CACHE_CONTROL))
         ->toEqual(Blitz::$plugin->settings->defaultCacheControlHeader);
 });
 
 test('Response contains the cache control header when the page is cacheable', function() {
     $response = sendRequest();
 
-    expect($response->headers->get('cache-control'))
+    expect($response->headers->get(HeaderEnum::CACHE_CONTROL))
         ->toEqual(Blitz::$plugin->settings->cacheControlHeader);
 });
 
@@ -47,25 +49,16 @@ test('Response contains the expired cache control header and the cache is refres
 
     $response = Blitz::$plugin->cacheRequest->getCachedResponse($siteUri);
 
-    expect($response->headers->get('cache-control'))
+    expect($response->headers->get(HeaderEnum::CACHE_CONTROL))
         ->toEqual(Blitz::$plugin->settings->cacheControlHeaderExpired);
 });
 
-test('Response adds the `X-Powered-By` header once', function() {
+test('Response adds the powered by header', function() {
     Craft::$app->config->general->sendPoweredByHeader = true;
     $response = sendRequest();
 
-    expect($response->headers->get('x-powered-by'))
-        ->toContainOnce('Blitz', 'Craft CMS');
-});
-
-test('Response overwrites the `X-Powered-By` header', function() {
-    Craft::$app->config->general->sendPoweredByHeader = false;
-    $response = sendRequest();
-
-    expect($response->headers->get('x-powered-by'))
-        ->toContainOnce('Blitz')
-        ->not()->toContain('Craft CMS');
+    expect($response->headers->get(HeaderEnum::X_POWERED_BY, first: false))
+        ->toContainEqual('Blitz');
 });
 
 test('Response contains output comments when enabled', function(bool|int $value) {
@@ -76,7 +69,7 @@ test('Response contains output comments when enabled', function(bool|int $value)
         ->toContain('Cached by Blitz');
 })->with([
     'true' => true,
-    'served' => SettingsModel::OUTPUT_COMMENTS_SERVED,
+    'SettingsModel::OUTPUT_COMMENTS_CACHED' => SettingsModel::OUTPUT_COMMENTS_CACHED,
 ]);
 
 test('Response does not contain output comments when disabled', function(bool|int $value) {
@@ -84,10 +77,10 @@ test('Response does not contain output comments when disabled', function(bool|in
     $response = sendRequest();
 
     expect($response->content)
-        ->not()->toContain('Served by Blitz on');
+        ->not()->toContain('Cached by Blitz');
 })->with([
     'false' => false,
-    'cached' => SettingsModel::OUTPUT_COMMENTS_CACHED,
+    'SettingsModel::OUTPUT_COMMENTS_SERVED' => SettingsModel::OUTPUT_COMMENTS_SERVED,
 ]);
 
 test('Response with mime type has headers and does not contain output comments', function() {
@@ -96,7 +89,7 @@ test('Response with mime type has headers and does not contain output comments',
     Blitz::$plugin->cacheStorage->save($output, $siteUri);
     $response = Blitz::$plugin->cacheRequest->getCachedResponse($siteUri);
 
-    expect($response->headers->get('Content-Type'))
+    expect($response->headers->get(HeaderEnum::CONTENT_TYPE))
         ->toBe('application/json')
         ->and($response->content)
         ->toBe($output);
@@ -107,25 +100,12 @@ test('Response is encoded when compression is enabled', function() {
     $siteUri = createSiteUri();
     Blitz::$plugin->cacheStorage->compressCachedValues = true;
     Blitz::$plugin->cacheStorage->save($output, $siteUri);
-    Craft::$app->getRequest()->headers->set('Accept-Encoding', 'deflate, gzip');
+    Craft::$app->getRequest()->headers->remove(HeaderEnum::ACCEPT_ENCODING);
+    Craft::$app->getRequest()->headers->set(HeaderEnum::ACCEPT_ENCODING, 'deflate, gzip');
     $response = Blitz::$plugin->cacheRequest->getCachedResponse($siteUri);
 
-    expect($response->headers->get('Content-Encoding'))
+    expect($response->headers->get(HeaderEnum::CONTENT_ENCODING))
         ->toBe('gzip')
         ->and(gzdecode($response->content))
         ->toBe($output);
-});
-
-test('Response is not encoded when compression is disabled', function() {
-    $output = createOutput();
-    $siteUri = createSiteUri();
-    Blitz::$plugin->cacheStorage->compressCachedValues = false;
-    Blitz::$plugin->cacheStorage->save($output, $siteUri);
-    Craft::$app->getRequest()->headers->set('Accept-Encoding', 'deflate, gzip');
-    $response = Blitz::$plugin->cacheRequest->getCachedResponse($siteUri);
-
-    expect($response->headers->get('Content-Encoding'))
-        ->toBeNull()
-        ->and($response->content)
-        ->toContain($output);
 });